Remove the file-based stuff
This script started as a way to copy around some files that I had set to
certain percentages using sed. But as it grew, I realized I could use jq
to do it and there would be no need for any files anywhere else.

So this commit removes all of that stuff and just uses the jq functions
to set everything.

There were some other changes and cleanup, but that's the main thing.

This commit is because I'm also about to remove the higher_chance and
decent_chance stuff since that can be done with set_current_chance,
which is how they're implemented anyway.