Initial check-in of GUI version of the script
As noted in the script, this was generated by ChatGPT with my help
regarding verbal prompts. Like:

"So far this is great. But can you implement the functionality to set all
maps to 100% if the `all` option is used for "Modify All Bosses for
Map". So if iterating over the list of valid maps, `all` should be
ignored since it's not an actual map but a representation of all maps."

or

"So I'd like to add a couple of things:
1. Back up the config file before saving the changes. Keeping last 10
   backups. I don't know if I want a configurable path location for so
   for now just use the same directory as the config file, perhaps in a
   'backups' subdirectory, ensuring it exists before writing to it.
2. How can I put some help text somewhere, preferably while hovering
   over the text label?"